Include refactor
Moved all includes to an include directory instead of putting them next
to their source files. This should make development in paradise cleaner
in the future as there is one source for all includes. Some minor
additions with this commit is renaming io.h to serial.io for additional
clarity.
