+++ include/paradise/processes.h
@@ -1,12 +1,12 @@
 #ifndef PROCESS_H_
 #define PROCESS_H_
 
-#include "ipc/mailbox.h"
-#include "lib/types.h"
-#include "memory/heap.h"
-#include "memory/mem.h"
+#include "heap.h"
+#include "mailbox.h"
+#include "mem.h"
 #include "queue.h"
 #include "rb_tree.h"
+#include "types.h"
 
 typedef struct {
     u32 page_dir_paddr;
