+++ include/paradise/serial.h
@@ -1,7 +1,7 @@
-#ifndef IO_H_
-#define IO_H_
+#ifndef SERIAL_H_
+#define SERIAL_H_
 
-#include "lib/types.h"
+#include "types.h"
 
 #define PORT_COM1 0x3F8
 
@@ -10,4 +10,4 @@ void serial_init();
 u8 serial_read();
 void serial_write(u8 data);
 
-#endif // IO_H_
+#endif // SERIAL_H_
