+++ include/paradise/strings.h
