+++ include/paradise/syscall.h
@@ -1,8 +1,8 @@
 #ifndef SYSCALL_H_
 #define SYSCALL_H_
 
-#include "lib/types.h"
-#include "process/processes.h"
+#include "processes.h"
+#include "types.h"
 
 typedef struct {
     u32 ret;
@@ -10,10 +10,14 @@ typedef struct {
 } SyscallResult;
 
 #define SYSCALL_RET(ret)                                                       \
-    return (SyscallResult) { ret, 0 }
+    return (SyscallResult) {                                                   \
+        ret, 0                                                                 \
+    }
 
 #define SYSCALL_ERR(err)                                                       \
-    return (SyscallResult) { 0, err }
+    return (SyscallResult) {                                                   \
+        0, err                                                                 \
+    }
 
 void register_syscall(u32 num, void *syscall);
 bool dispatch_syscall(CpuContext *ctx);
