+++ include/paradise/testing.h
@@ -1,7 +1,7 @@
 #ifndef TESTING_H_
 #define TESTING_H_
 
-#include "boot/multiboot.h"
+#include "multiboot.h"
 
 void kernel_test();
