+++ include/paradise/util.h
