+++ src/lib/libp.c
@@ -1,4 +1,4 @@
-#include "libp.h"
+#include <paradise/libp.h>
 
 // NOTE: we can optimize our mass memory operations by operating on the 4-byte
 // aligned u32 level instead of at a byte level
